[BUGFIX] Raise index_rel.first to int(11)
The index_rel.first database table of indexed_search
is used to store the number of occurrences of words,
which could be more in bigger indices.

Resolves: #75406
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/47529
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/indexed_search/ext_tables.sql
@@ -49,7 +49,7 @@ CREATE TABLE index_rel (
   phash int(11) DEFAULT '0' NOT NULL,
   wid int(11) DEFAULT '0' NOT NULL,
   count tinyint(3) unsigned DEFAULT '0' NOT NULL,
-  first tinyint(3) unsigned DEFAULT '0' NOT NULL,
+  first int(11) unsigned DEFAULT '0' NOT NULL,
   freq smallint(5) unsigned DEFAULT '0' NOT NULL,
   flags tinyint(3) unsigned DEFAULT '0' NOT NULL,
   PRIMARY KEY (phash,wid),
